Reduce code size and some updates (#75)
* Remove unused babel plugin

* Remove bloat to reduce code size

With some of the newer additions to the script, the repo became slow to
load. This is due to us presenting the script's code for users to copy in the
main `Readme.md` file, and this script contains ~52k characters, which
is a significant number. In order to speed up the repo's load-time a
bit, this commit does the following:

1. Remove unused `uniqueMerge` method. Surprisingly remove >1k
   characters (probably related to the usage of `Set`).

2. Remove `storage.ts` abstraction class for local/session storage
   methods. Doesn't abstract that much and we barely use it ATM.

3. Add `.htaccess` file with dedicated logic regarding gzip compression.
   This addition should tell GH which files to compress in the repo.

* Update logo style

* Update screenshots in readme

* Enlarge progressbar to make it more noticable